+++ src/Commands/ListTranslationKeysCommand.php
@@ -5,7 +5,6 @@ namespace PHPUnuhi\Commands;
 use PHPUnuhi\Configuration\ConfigurationLoader;
 use PHPUnuhi\Exceptions\ConfigurationException;
 use PHPUnuhi\Services\Loaders\Xml\XmlLoader;
-use PHPUnuhi\Traits\CommandOutputTrait;
 use PHPUnuhi\Traits\CommandTrait;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -13,10 +12,9 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
-class ListTranslationsCommand extends Command
+class ListTranslationKeysCommand extends Command
 {
     use CommandTrait;
-    use CommandOutputTrait;
 
     /**
      * @return void
@@ -24,7 +22,7 @@ class ListTranslationsCommand extends Command
     protected function configure()
     {
         $this
-            ->setName(CommandNames::LIST_TRANSLATIONS)
+            ->setName(CommandNames::LIST_TRANSLATION_KEYS)
             ->setDescription('')
             ->addOption('configuration', null, InputOption::VALUE_REQUIRED, '', '');
 
@@ -57,12 +55,8 @@ class ListTranslationsCommand extends Command
         foreach ($config->getTranslationSets() as $set) {
             $io->section('Translation Set: ' . $set->getName());
 
-            foreach ($set->getLocales() as $locale) {
-                $rows = [];
-                foreach ($locale->getTranslations() as $translation) {
-                    $rows[] = [$locale->getName(), $translation->getKey(), $translation->getValue()];
-                }
-                $this->showTranslationTable($rows, $io);
+            foreach ($set->getAllTranslationIDs() as $id) {
+                $io->writeln('   [~] ' . $id);
             }
         }